refactoring conversationalClientResolver
