+++ src/app/smartRecom/services/aiService.ts
@@ -2,24 +2,11 @@
 import { GoogleGenAI } from '@google/genai';
 import { FileData } from "../models/types";
 
-export async function summarizeMultipleReports(reports: FileData[]): Promise<string> {
-  const summaries = reports.map(r => r.report.summary).join("\n");
-
-  const response = await askAI({
-    systemPrompt: "Eres un analista de soporte al cliente. Tu trabajo es escribir un resumen narrativo breve y claro basado en varios reportes de llamadas de un mismo agente.",
-    context: summaries,
-    question: `Genera un resumen narrativo en lenguaje natural que combine los puntos clave de todos los reportes anteriores. Este resumen será mostrado a un agente humano para entender el contexto emocional y temático del cliente.`,
-  });
-
-  return response;
-}
-
 interface AIInput {
   systemPrompt: string;
   context: string;
   question: string;
 }
-
 const ai = new GoogleGenAI({
   apiKey: process.env.GEMINI_API_KEY || '',
 });
@@ -45,4 +32,17 @@ export async function askAI({ systemPrompt, context, question }: AIInput): Promi
     console.error('Error en Gemini API:', error);
     return 'No se pudo generar una recomendación. Intenta más tarde.';
   }
+}
+
+
+export async function summarizeMultipleReports(reports: string): Promise<string> {
+  const summaries = reports;
+
+  const response = await askAI({
+    systemPrompt: "Eres un analista de soporte al cliente. Tu trabajo es escribir un resumen narrativo breve y claro de dos párrafos, basado en varios reportes de llamadas de un mismo agente.",
+    context: summaries,
+    question: `Genera un resumen narrativo en lenguaje natural que combine los puntos clave de todos los reportes anteriores. Este resumen será mostrado a un agente humano para entender el contexto emocional y temático del cliente.`,
+  });
+
+  return response;
 }
