+++ src/app/smartRecom/services/conversationalClientResolver.ts
@@ -1,12 +1,9 @@
+// 📁 core/chatClientController.ts
 import Fuse from "fuse.js";
 import { askAI } from "./aiService";
-import { mockClients } from "../data/mockClientDB";
-import { api } from "@/trpc/react";
 import { client } from "@prisma/client";
 
 let lastSuggestedClient: { id: number; name: string } | null = null;
-
-
 
 /**
  * Controlador principal del chat IA para identificar clientes y responder según el flujo natural.
@@ -29,7 +26,7 @@ export async function chatClientController(message: string, clients: client[]): 
     const confirm = await askAI({
       systemPrompt: "Eres un validador. Responde solo 'yes' o 'no'.",
       context: "",
-      question: `El usuario respondió: "${message}". ¿Es una afirmación clara de que acepta el nombre sugerido?`,
+      question: `El usuario respondió: \"${message}\". ¿Es una afirmación clara de que acepta el nombre sugerido?`,
     });
 
     if (confirm.toLowerCase().includes("yes")) {
@@ -37,13 +34,6 @@ export async function chatClientController(message: string, clients: client[]): 
       const clientName = lastSuggestedClient.name;
       lastSuggestedClient = null;
 
-      // Log para consola
-      console.log("AI resolver result:\n" + JSON.stringify({
-        type: "client",
-        id: clientId,
-      }, null, 2));
-
-      // JSON para el frontend
       return JSON.stringify({
         type: "client",
         id: clientId,
@@ -55,42 +45,35 @@ export async function chatClientController(message: string, clients: client[]): 
     }
   }
 
-  // Paso 2: IA interpreta a quién se refiere el mensaje
-  // const clientNames = mockClients.map((c) => c.name).join(", ");
-
-  const clientNames = clients?.map((c) => (c.firstname + " " + c.lastname)).join(", ");
-  const inferredName = await askAI({
-    systemPrompt: "Eres un asistente que identifica a qué cliente se refiere un mensaje.",
+  // Paso 2: IA interpreta si el mensaje es una solicitud de info + nombre
+  const clientNames = clients.map(c => `${c.firstname} ${c.lastname}`).join(", ");
+  const instructionCheck = await askAI({
+    systemPrompt: "Eres un asistente que detecta si el usuario quiere información sobre un cliente.",
     context: `Nombres disponibles: ${clientNames}`,
-    question: `¿A qué nombre completo se refiere este mensaje?: "${message}". Si no puedes estar seguro, devuelve null.`,
+    question: `El usuario dijo: \"${message}\". ¿Está pidiendo información de un cliente con nombre incluido? Si sí, responde solo el nombre. Si no, responde 'null'.`,
   });
 
-  // console.log("Inferred name:", inferredName);
-  const match = fuse.search(inferredName)?.[0]?.item;
-  // console.log("Match:", match);
+  if (!instructionCheck || instructionCheck.toLowerCase() === "null") {
+    return "No entendí a qué cliente te refieres o qué necesitas. Por favor, intenta escribir algo como: 'dame info de Juan Pérez'.";
+  }
+
+  const match = fuse.search(instructionCheck)?.[0]?.item;
 
   if (!match) {
     return "No pude identificar a ningún cliente con base en ese mensaje. ¿Puedes reformularlo?";
   }
 
-  // Paso 3: Si no hay coincidencia exacta, sugerir nombre
-  if (!inferredName.toLowerCase().includes(match.name.toLowerCase())) {
+  if (!instructionCheck.toLowerCase().includes(match.name.toLowerCase())) {
     lastSuggestedClient = { id: match.id, name: match.name };
 
     const suggestionReply = await askAI({
       systemPrompt: "Eres un asistente que confirma sugerencias de nombres.",
-      context: `El usuario dijo: "${message}". El nombre más parecido es: "${match.name}"`,
+      context: `El usuario dijo: \"${message}\". El nombre más parecido es: \"${match.name}\"` ,
       question: "Redacta una respuesta natural preguntando si se refiere a ese nombre.",
     });
 
     return suggestionReply;
   }
-
-  // Coincidencia clara
-  console.log("AI resolver result:\n" + JSON.stringify({
-    type: "client",
-    id: match.id,
-  }, null, 2));
 
   return JSON.stringify({
     type: "client",
