+++ src/app/smartRecom/services/openR+oldstuff/aiServiceOpenRouter.ts
