+++ src/app/smartRecom/services/openR+oldstuff/smartInsights.ts
